large fileに対応

+++ encoder.h
@@ -32,7 +32,7 @@
 
 typedef struct EncodeDictionary
 {
-  uint txt_len;
+  uint64 txt_len;
   uint seq_len;
   uint num_rules;
   uint num_usedrules;
